# out2/src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: alberto.est.jimenez@gmail.com
    name: Alberto Jimenez
  description: Documentation for the client-accessible endpoints of the StackV orchestrator.
  title: StackV Orchestrator Developer API
  version: 0.8.2
servers:
- url: https://stackv.dev3.virnao.com/StackV-web/restapi/
...
paths:
  /instance:
    get:
      description: Retrieves UUID for new instance from backend.
      responses:
        "200":
          content:
            application/json:
              schema:
                example: 3b7d41e1-b37f-4c64-84d2-d307b5e47370
                type: string
          description: Generated UUID.
        "500":
          $ref: "#/components/responses/Error"
      summary: Generate random UUID
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{search}:
    get:
      description: Retrieves the full status and details of a specified instance.
      parameters:
      - description: Service instance UUID or alias.
        example:
        - ae4d45d0-4420-42d6-8641-0a1012d47c99
        - DNC OpenNSA 21
        explode: false
        in: path
        name: search
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Root_Schema"
          description: Full instance details object.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get instance details.
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
    post:
      description: Creates a new service instance with the given UUID and intent specification.
      parameters:
      - description: Service instance UUID.
        explode: false
        in: path
        name: search
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
        description: Service instance creation request object.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                example: ae4d45d0-4420-42d6-8641-0a1012d47c99
                type: string
          description: "Service instance UUID, in case it was generated during workflow."
        "500":
          $ref: "#/components/responses/Error"
      summary: Create new service instance
      tags:
      - instance
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{search}/status:
    get:
      description: Retrieves the full instance status.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: search
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                example: Create - FAILED
                type: string
          description: Concatenated instance superstate and substate.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get instance status
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/substatus:
    get:
      description: Retrieves only the backend service status.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                description: The status of an instance as it proceeds through an orchestration
                  phase.
                enum:
                - INIT
                - COMPILED
                - PROPAGATED
                - COMMITTING
                - COMMITTED
                - READY
                - FAILED
                - ERROR
                type: string
          description: Workflow status.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get instance substatus
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/status/phase:
    get:
      description: Retrieves the instance's orchestration phase.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1instance~1%7BsiUUID%7D~1superstate~1%7Bstate%7D/put/parameters/1/schema"
          description: Orchestration status.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get instance phase
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
      - tag: admin
  /instance/{siUUID}/status/configuration:
    get:
      description: Retrieves the instance's configuration status.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                description: The status of whether an instance is actively provisioned
                  or configured.
                enum:
                - STABLE
                - SCHEDULED
                - EXPIRED
                - PENDING
                - UNSTABLE
                - UNKNOWN
                type: string
          description: Configuration status.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get instance configuration
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/superstate/{state}:
    put:
      description: Updates the orchestration phase of the specified service instance.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      - description: New orchestration phase.
        example: CANCEL
        explode: false
        in: path
        name: state
        required: true
        schema:
          description: The overarching phase of an instance's lifecycle.
          enum:
          - CREATE
          - CANCEL
          - REINSTATE
          - MODIFY
          type: string
        style: simple
      responses:
        "204":
          description: No content response.
        "500":
          $ref: "#/components/responses/Error"
      summary: Change instance orchestration phase
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/{action}:
    put:
      description: Commit an operation for specified service instance.
      parameters:
      - description: service instance UUID
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      - description: |-
          Service operation requested
           * `cancel` - Composite service to cancel, propagate, commit and verify instance.
           * `release` - Composite service to cancel and propagate in 2-phase commit process.
           * `verify` - Begin service verification process.
        explode: false
        in: path
        name: action
        required: true
        schema:
          $ref: "#/components/schemas/Operation"
        style: simple
      - description: Whether to commit the operation in a synchronous/blocking mode.
        explode: true
        in: query
        name: sync
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: "Whether to ensure service verifiction is run, even in the case\
          \ of failure states."
        explode: true
        in: query
        name: verify
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Whether to commit the operations in force mode.
        explode: true
        in: query
        name: force
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Whether to automatically proceed through propagate and commit.
        explode: true
        in: query
        name: proceed
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Explicit intent specified for certain operations.
        explode: true
        in: query
        name: intent
        required: false
        schema:
          type: string
        style: form
      responses:
        "204":
          description: No content response.
        "500":
          $ref: "#/components/responses/Error"
      summary: Operate on a service instance
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/modify:
    post:
      description: "Composite service to modify/recompile, propagate, commit and verify\
        \ (will create a new intent)"
      parameters:
      - description: service instance UUID
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
        description: Service instance creation request object.
        required: true
      responses:
        "204":
          description: No content response.
        "500":
          $ref: "#/components/responses/Error"
      summary: Modify a service instance
      tags:
      - instance
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/modifiable/{mode}:
    put:
      description: Enables or disables the modifiable option for the instance.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      - description: New value.
        explode: false
        in: path
        name: mode
        required: true
        schema:
          type: boolean
        style: simple
      responses:
        "204":
          description: Successful set.
        "500":
          $ref: "#/components/responses/Error"
      summary: Set instance modifiable flag
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
  /instance/{siUUID}/last/{state}:
    put:
      description: Overwrites the last state of the instance.
      parameters:
      - description: Service instance UUID.
        example: ae4d45d0-4420-42d6-8641-0a1012d47c99
        explode: false
        in: path
        name: siUUID
        required: true
        schema:
          type: string
        style: simple
      - description: New state.
        explode: false
        in: path
        name: state
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Successful set.
        "500":
          $ref: "#/components/responses/Error"
      summary: Set instance last state
      tags:
      - instance
      x-accepts:
      - application/json
      x-tags:
      - tag: instance
      - tag: admin
...
components:
  responses:
    Error:
      description: Internal server fault or standardized error response.
...
  schemas:
...
    Operation:
      enum:
      - cancel
      - force_cancel
      - release
      - force_release
      - provision
      - reinstate
      - rollback
      - force_retry
      - delete
      - force_delete
      - archive
      - unarchive
      - reset
      - verify
      - unverify
      type: string
...
  securitySchemes:
    OAuth2:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/auth
          scopes:
            openid: OIDC Mode.
          tokenUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/token
      type: oauth2